Changed: simpler OIDC config init
